Add C++ function to compute polynomial expansions
This function works row-by-row, so that we never have to store the
entire N x polynomial-terms(k) matrix in memory at once.  It is not
meant to be called directly by the user and will probably end up in a
header file in the final rewritten version of the package.