Rewrite bootstrap and fitting functions
Unfortunately, this has to be done as an omni-commit due to
chicken-and-egg problems.  The backend functions `fitAdaptiveLASSO()`
and `fitSCAD()` were rewritten in anticipation of the additional
arguments to `polywog()`; either would have been broken without the
other.  Along the way, I also did a bunch of cleanup (closes #8,
incidentally), cut down on code rewriting, only compute the
polynomial-expanded model matrix when necessary, etc., hence the large
amount of file-by-file changes.

This commit also wipes out the now-defunct `polym2()` and `rawpoly()`
functions, all of which are implemented better (and more readably) in
C++ code.

Changes to the bootstrap are also implemented here since `bootPolywog()`
is tied so closely to `polywog()`.  However, this commit does break
`cv.polywog()`.