Change default thresh/maxit in bootstrapping
Previously, the default was to use the same default as in `polywog()`,
depending on the penalization method used.  But that required clunky
code like

```r
m1 <- polywog(y ~ x, thresh = 1e-10,
              control.boot = control.bp(thresh = 1e-10))
```

in order to make the bootstrap behave the same as original fitting.
The new behavior (which requires adding `thresh` and `maxit` elements to
`"polywog"` objects) makes this easier.  (As a bonus, it also results in
more aesthetically pleasing code in the examples, where relatively high
thresholds have to be used in order not to run afoul of CRAN's
computation time requirements.)